+++ src/ir_MagiQuest.hpp
@@ -14,7 +14,7 @@
  ************************************************************************************
  * MIT License
  *
- * Copyright (c) 2017-2023 E. Stuart Hicks <ehicks@binarymagi.com>, Armin Joachimsmeyer
+ * Copyright (c) 2017-2024 E. Stuart Hicks <ehicks@binarymagi.com>, Armin Joachimsmeyer
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -79,6 +79,8 @@
  // Checksum (+ sum of the 5 bytes before == 0)
  + 250,- 900 + 300,- 900 + 250,- 850 + 550,- 600
  + 600,- 550 + 300,- 900 + 250,- 850 + 550
+
+ // No stop bit!
  */
 // MSB first, 8 start bits (zero), 31 wand id bits, 9 magnitude bits 8 checksum bits and no stop bit => 56 bits
 #define MAGIQUEST_CHECKSUM_BITS     8   // magiquest_t.cmd.checksum
@@ -108,7 +110,8 @@
 
 // assume 110 as repeat period
 struct PulseDistanceWidthProtocolConstants MagiQuestProtocolConstants = { MAGIQUEST, 38, MAGIQUEST_ZERO_MARK, MAGIQUEST_ZERO_SPACE,
-MAGIQUEST_ONE_MARK, MAGIQUEST_ONE_SPACE, MAGIQUEST_ZERO_MARK, MAGIQUEST_ZERO_SPACE, PROTOCOL_IS_MSB_FIRST, 110, NULL };
+MAGIQUEST_ONE_MARK, MAGIQUEST_ONE_SPACE, MAGIQUEST_ZERO_MARK, MAGIQUEST_ZERO_SPACE, PROTOCOL_IS_MSB_FIRST | SUPPRESS_STOP_BIT, 110,
+        NULL };
 //+=============================================================================
 //
 /**