+++ src/private/IRTimer.hpp
@@ -1450,26 +1450,26 @@ void timerEnableReceiveInterrupt() {
     timerAlarmEnable(s50usTimer);
 }
 
-#if !defined(ESP_ARDUINO_VERSION)
+#  if !defined(ESP_ARDUINO_VERSION)
 #define ESP_ARDUINO_VERSION 0
-#endif
-#if !defined(ESP_ARDUINO_VERSION_VAL)
+#  endif
+#  if !defined(ESP_ARDUINO_VERSION_VAL)
 #define ESP_ARDUINO_VERSION_VAL(major, minor, patch) 202
-#endif
-#if ESP_ARDUINO_VERSION < ESP_ARDUINO_VERSION_VAL(2, 0, 2)
+#  endif
+#  if ESP_ARDUINO_VERSION < ESP_ARDUINO_VERSION_VAL(2, 0, 2)
 void timerDisableReceiveInterrupt() {
     if (s50usTimer != NULL) {
         timerDetachInterrupt(s50usTimer);
         timerEnd(s50usTimer);
     }
 }
-#else
+#  else
 void timerDisableReceiveInterrupt() {
     if (s50usTimer != NULL) {
         timerAlarmDisable(s50usTimer);
     }
 }
-#endif
+#  endif
 
 // Undefine ISR, because we register/call the plain function IRReceiveTimerInterruptHandler()
 #  if defined(ISR)